Changed advanced menu function clearing local songs
Updated the function to clean out local songs on the hard drive without
removing files from the cloud, local dirty files, OR local files that
have not been uploaded yet.

Provides a way to clean house after the song manager gets big over time.

+++ src/menus/menu_advanced.py
@@ -1,13 +1,15 @@
+from src.Hash import Hash
 from src.TERMGUI.Log import Log
 from src.TERMGUI.Menu import Menu
 from src.TERMGUI.Dialog import Dialog
+from src.FileManagement.File import File
 from src.FileManagement.Folder import Folder
 from src.Project.Index import Index as ProjectIndex
 
 def menu_advanced():
     options = [
-        ["Clean Out Cache",           clean_out_cache],
-        ["Remove Extracted Projects", remove_extracted_projects],
+        ["Clean Out Cache",          clean_out_cache],
+        ["Clear All Local Projects", clear_all_local_projects],
     ]
 
     menu = Menu(
@@ -44,36 +46,35 @@ def clean_out_cache():
 
     return True
 
-def remove_extracted_projects():
-    dialog = Dialog(
-        title = "Remove Extracted Projects",
-        body  = [
-            f'This will remove all extracted projects from your local computer!',
-            f'\n', f'\n',
-            f'You have two options:',
-            f'\n', f'\n',
-            f'  - "all":    Remove ALL extracted projects', f'\n',
-            f'  - "remote": Remove ONLY the songs that are backed up on the drive',
-            f'\n', f'\n',
-            f'The "remote" option will NOT delete any un-uploaded projects you have saved,',
-            f'nor will either option touch ANY projects on the cloud.',
-            f'\n', f'\n',
+def clear_all_local_projects():
+    result = Dialog(
+        "Clear Local Projects",
+        [
+            f'This will clear all locally cached projects to save',
+            f'space on your hard drive.', f'\n',
+            f'This will NOT:', f'\n',
+            f'  - Delete any songs on the cloud', f'\n',
+            f'  - Delete any dirty projects on your computer', f'\n',
+            f'    that have not been uploaded yet', f'\n',
+            f'  - Remove any projects that have never been', f'\n',
+            f'    uploaded to the cloud yet', f'\n',
         ]
-    )
+    ).confirm()
 
-    ans = dialog.get_mult_choice(["all","remote","back"])
+    if result:
+        for project in ProjectIndex.get_all_projects():
+            if not project.is_dirty() and project.is_remote():
+                # Just in case we have it locked
+                project.remove_lock()
 
-    if ans == "back":
-        return True
+                # Remove extracted folder and cached file
+                File.delete(project.get_cache_file())
+                Folder.delete(project.get_root_dir())
 
-    if ans == "all":
-        Folder.clear(f'extracted_songs')
-    elif ans == "remote":
-        projects = ProjectIndex.get_local_projects()
-        for project in projects:
-            if project.is_remote():
-                Folder.delete( project.get_root_dir() )
+                # Remove hash from local db
+                Hash.remove_project_hash(project)
 
-    Menu.notice = "Extracted Projects Cleaned Out!"
-
+            Menu.notice = f'Local projects cleared!'
+    else:
+        Menu.notice = f'No local projects were cleared..'
     return True
